Fix nested navigation parsing
Sometimes, <ol> doesn't necessarily contain an <a> tag as a direct
child.
Thus we need to check the direct children's tagname, rather than use
query selector which might go deeper than 1 level.

When this happens, we should parse the <ol> subtree, but at the same
level of nesting and preserve the current order.